[Files]: Switch FileIndexService to use SQL storage (Step 3 of 3)
Switches to persistent storage.

Bug: b:327534824
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5507696
Commit-Queue: Bo Majewski <majewski@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1296709}

+++ chrome/browser/ash/file_manager/indexing/file_index_service.cc
@@ -6,7 +6,7 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 #include "chrome/browser/ash/file_manager/indexing/file_index_service.h"
 
 #include "chrome/browser/ash/file_manager/indexing/file_index_impl.h"
-#include "chrome/browser/ash/file_manager/indexing/ram_storage.h"
+#include "chrome/browser/ash/file_manager/indexing/sql_storage.h"
 
 namespace file_manager {
 
@@ -33,30 +33,65 @@ namespace file_manager {
 //               |               |
 //        [ RamStorage ]   [ SqlStorage ]
 
+namespace {
+
+base::FilePath MakeDbPath(const Profile* const profile) {
+  return profile->GetPath()
+      .AppendASCII("file_manager")
+      .AppendASCII("file_index.db");
+}
+
+constexpr char kSqlDatabaseUmaTag[] =
+    "FileBrowser.FileIndex.SqlDatabase.Status";
+
+}  // namespace
+
 FileIndexService::FileIndexService(Profile* profile)
-    : file_index_impl_(
-          std::make_unique<FileIndexImpl>(std::make_unique<RamStorage>())) {
-  DCHECK(profile);
+    : file_index_impl_(std::make_unique<FileIndexImpl>(
+          std::make_unique<SqlStorage>(MakeDbPath(profile),
+                                       kSqlDatabaseUmaTag))) {}
+
+bool FileIndexService::Init() {
+  if (inited_) {
+    return true;
+  }
+  if (!file_index_impl_->Init()) {
+    return false;
+  }
+  inited_ = true;
+  return true;
 }
 
 FileIndexService::~FileIndexService() = default;
 
 OpResults FileIndexService::UpdateFile(const std::vector<Term>& terms,
                                        const FileInfo& info) {
+  if (!inited_) {
+    return kUninitialized;
+  }
   return file_index_impl_->UpdateFile(terms, info);
 }
 
 OpResults FileIndexService::AugmentFile(const std::vector<Term>& terms,
                                         const FileInfo& info) {
+  if (!inited_) {
+    return kUninitialized;
+  }
   return file_index_impl_->AugmentFile(terms, info);
 }
 
 OpResults FileIndexService::RemoveFile(const GURL& url) {
+  if (!inited_) {
+    return kUninitialized;
+  }
   return file_index_impl_->RemoveFile(url);
 }
 
 // Searches the index for file info matching the specified query.
 SearchResults FileIndexService::Search(const Query& query) {
+  if (!inited_) {
+    return SearchResults();
+  }
   return file_index_impl_->Search(query);
 }
